exzoo_backend/pyproject.toml
[tool.poetry]
name = "exzoo-backend"
version = "0.1.0"
description = ""
authors = ["Denis Tokarev <wf.denis.tokarev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
psycopg2 = "^2.9.9"
django-debug-toolbar = "^4.4.6"
pillow = "^10.4.0"
djoser = "^2.2.3"
django-filter = "^24.3"
drf-yasg = "^1.21.8"
mypy = "^1.13.0"
django-stubs = "^5.1.1"
djangorestframework-stubs = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
